Refactored game files into gamification folder + automatic delete items from inventory object if not found in itemData
