+++ public/game-overlays/boatGrow/boatGrow.js
@@ -1,4 +1,4 @@
-import { getRandomFishByCost } from "../../src/ItemData.ts";
+import { getRandomFishByCost } from "../../src/gamification/ItemData.ts";
 
 // Global config + state
 let energyCost = 0;
